src/bonnet/requirements.txt
board
adafruit-blinka
adafruit-circuitpython-ssd1306
Pillow
psutil